build: Don't do the full wheel build to check the docs syntax.
Also drop the changelog from the long description.  The way it is right
now, that will never succeed the strict checks becaause of the way
setuptools currently parses the RST.  It generates implicit labels for
all headings by default and then it complains when there are duplicates.

This will always happen with changelogs because the sub-headings under
a changelog entry will always be one of a small set of values (added,
fixed, etc.)